KEGGENZYMEID2GO {KEGG.db}	R Documentation

An annotation data object that maps Enzyme Commission numbers to Gene Ontology identifiers

Description

KEGGENZYMEID2GO maps Enzyme Commission numbers to Gene Ontoloty (GO) identifiers

Details

This is an R object containing key and value pairs. Keys are Enzyme Commission numbers and values are GO identifiers. Values are vectors of length 1. Enzyme Commission numbers that can not be mapped to a GO term are assigned a value NA.

Mappings were based on data provided by: Gene Ontology External Link http://www.geneontology.org/external2go With a date stamp from the source of: 2015-Sepec2go27

References

ftp://ftp.genome.ad.jp/pub/kegg/pathways

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
xx <- as.list(KEGGENZYMEID2GO)
if(length(xx) > 0){
    # Get the value of the first key
    xx[[1]]
    # Get values for a few keys
    if(length(xx) >= 3)
        xx[1:3]
    }

[Package KEGG.db version 3.2.3 Index]