KEGGGO2ENZYMEID {KEGG.db}	R Documentation

An annotation data object that maps Gene Ontology (GO) identifiers to Enzyme Commission numbers

Description

KEGGGO2ENZYMEID maps GO identifiers to Enzyme Commission numbers

Details

This is an R object containing key and value pairs. Keys are GO identifiers and values are Enzyme Commission numbers. Values are vectors of length 1. GO identifiers can not be mapped to any Enzyme Commission number are assigned NAs.

Mappings are based on data provided by: Gene Ontology External Link http://www.geneontology.org/external2go With a date stamp from the source of: 2015-Sepec2go27

References

ftp://ftp.genome.ad.jp/pub/kegg/pathways

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
xx <- as.list(KEGGGO2ENZYMEID)
if(length(xx) > 0){
    # Get the value of the first key
    xx[[1]]
    # Get values for a few keys
    if(length(xx) >= 3){
    	xx[1:3]
    }
}

[Package KEGG.db version 3.2.3 Index]