KEGG_dbconn {KEGG.db}	R Documentation

Collect information about the package annotation DB

Description

Some convenience functions for getting a connection object to (or collecting information about) the package annotation DB.

Usage

  KEGG_dbconn()
  KEGG_dbfile()
  KEGG_dbschema(file="", show.indices=FALSE)
  KEGG_dbInfo()

Arguments

file
A connection, or a character string naming the file to print to (see the file argument of the cat function for the details).

show.indices
The CREATE INDEX statements are not shown by default. Use show.indices=TRUE to get them.

Details

KEGG_dbconn returns a connection object to the package annotation DB. IMPORTANT: Don't call dbDisconnect on the connection object returned by KEGG_dbconn or you will break all the AnnDbObj objects defined in this package!

KEGG_dbfile returns the path (character string) to the package annotation DB (this is an SQLite file).

KEGG_dbschema prints the schema definition of the package annotation DB.

KEGG_dbInfo prints other information about the package annotation DB.

Value

KEGG_dbconn: a DBIConnection object representing an open connection to the package annotation DB.

KEGG_dbfile: a character string with the path to the package annotation DB.

KEGG_dbschema: none (invisible NULL).

KEGG_dbInfo: none (invisible NULL).

See Also

dbGetQuery, dbConnect, dbconn, dbfile, dbschema, dbInfo

Examples

  library(DBI)
  ## Count the number of rows in the "pathway2name" table:
  dbGetQuery(KEGG_dbconn(), "SELECT COUNT(*) FROM pathway2name")

  KEGG_dbschema()

  KEGG_dbInfo()

[Package KEGG.db version 3.2.3 Index]